+++ src/main/java/vectorwing/farmersdelight/items/Foods.java
@@ -126,7 +126,7 @@ public class Foods
 			.effect(() -> new EffectInstance(ModEffects.COMFORT.get(), 2400, 0), 1.0F).build();
 	public static final Food FRIED_RICE = (new Food.Builder())
 			.nutrition(10).saturationMod(0.8f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 2400, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 2400, 0), 1.0F).build();
 	public static final Food PUMPKIN_SOUP = (new Food.Builder())
 			.nutrition(10).saturationMod(0.9f)
 			.effect(() -> new EffectInstance(ModEffects.COMFORT.get(), 2400, 0), 1.0F).build();
@@ -140,45 +140,45 @@ public class Foods
 	// Plated Foods
 	public static final Food BACON_AND_EGGS = (new Food.Builder())
 			.nutrition(9).saturationMod(0.6f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 1200, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 1200, 0), 1.0F).build();
 	public static final Food STEAK_AND_POTATOES = (new Food.Builder())
 			.nutrition(12).saturationMod(0.8f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 2400, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 2400, 0), 1.0F).build();
 	public static final Food PASTA_WITH_MEATBALLS = (new Food.Builder())
 			.nutrition(12).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food PASTA_WITH_MUTTON_CHOP = (new Food.Builder())
 			.nutrition(12).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food ROASTED_MUTTON_CHOPS = (new Food.Builder())
 			.nutrition(12).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food VEGETABLE_NOODLES = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food RATATOUILLE = (new Food.Builder())
 			.nutrition(9).saturationMod(0.6f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 1200, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 1200, 0), 1.0F).build();
 	public static final Food SQUID_INK_PASTA = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food GRILLED_SALMON = (new Food.Builder())
 			.nutrition(12).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 
 	// Feast Portions
 	public static final Food ROAST_CHICKEN = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food STUFFED_PUMPKIN = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food HONEY_GLAZED_HAM = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 	public static final Food SHEPHERDS_PIE = (new Food.Builder())
 			.nutrition(14).saturationMod(0.9f)
-			.effect(() -> new EffectInstance(ModEffects.NOURISHED.get(), 4800, 0), 1.0F).build();
+			.effect(() -> new EffectInstance(ModEffects.NOURISHMENT.get(), 4800, 0), 1.0F).build();
 
 	public static final Food DOG_FOOD = (new Food.Builder())
 			.nutrition(4).saturationMod(0.2f).meat().build();