+++ src/main/java/vectorwing/farmersdelight/items/MealItem.java
@@ -1,0 +1,28 @@
+package vectorwing.farmersdelight.items;
+
+import net.minecraft.client.util.ITooltipFlag;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.world.World;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import vectorwing.farmersdelight.setup.Configuration;
+import vectorwing.farmersdelight.utils.TextUtils;
+
+import javax.annotation.Nullable;
+import java.util.List;
+
+public class MealItem extends ConsumableItem
+{
+	public MealItem(Properties properties) {
+		super(properties);
+	}
+
+	@Override
+	@OnlyIn(Dist.CLIENT)
+	public void appendHoverText(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+		if (Configuration.MEAL_EFFECT_TOOLTIP.get()) {
+			TextUtils.addFoodEffectTooltip(stack, tooltip, 1.0F);
+		}
+	}
+}