+++ src/main/java/vectorwing/farmersdelight/registry/ModEffects.java
@@ -12,6 +12,6 @@ public class ModEffects
 {
 	public static final DeferredRegister<Effect> EFFECTS = DeferredRegister.create(ForgeRegistries.POTIONS, FarmersDelight.MODID);
 
-	public static final RegistryObject<Effect> NOURISHED = EFFECTS.register("nourished", NourishmentEffect::new);
+	public static final RegistryObject<Effect> NOURISHMENT = EFFECTS.register("nourished", NourishmentEffect::new);
 	public static final RegistryObject<Effect> COMFORT = EFFECTS.register("comfort", ComfortEffect::new);
 }